refactor: move all icon components files into icons subfolder
