+++ src/components/tools/icons/languages.js
